compose vehicle supervisor. split into .hpp. inheriting supervisor for launching version
