+++ src/hardware/vehicle_supervisor/src/node_supervisor_launch.cpp
@@ -1,11 +1,11 @@
-#include "component_supervisor.hpp"
+#include "component_supervisor_launch.hpp"
 #include "rclcpp/rclcpp.hpp"
 
 int main(int argc, char *argv[]) {
     rclcpp::init(argc, argv);
 
     rclcpp::NodeOptions options;
-    auto node = std::make_shared<vehicle_supervisor::ASSupervisor>(options);
+    auto node = std::make_shared<vehicle_supervisor::ASSupervisorLaunch>(options);
 
     rclcpp::executors::MultiThreadedExecutor executor;
     executor.add_node(node);
